Update github actions
Following the description here:

https://niteo.co/blog/staging-like-its-2020

(Which has errors if copied directly). This expands the simple
review-app-test setup with a view to adding cypress tests after
it.

GITHUB_TOKEN and GITHUB_REPO variables need to be added to the
pipeline config vars on Heroku.

A separate release script is added and called from the Procfile.
It runs the previous npm commands and also performs a
"repository_dispatch" call to github to kick off the action.